+++ CHANGELOG.md
@@ -6,6 +6,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.8.0] - 2026-08-05
+
 ### Added
 
 - **`native_apps: [loopback_include_localhost: true]`**, an opt-in that widens